integrate MovementBridge service
