+++ protocol-units/bridge/chains/movement/src/lib.rs
@@ -1,319 +1,70 @@
-use crate::utils::MovementAddress;
-use anyhow::Result;
-use aptos_sdk::{
-	move_types::language_storage::TypeTag,
-	rest_client::{Client, FaucetClient},
-	types::LocalAccount,
-};
-use aptos_types::account_address::AccountAddress;
+use std::collections::HashMap;
+
 use bridge_shared::{
-	bridge_contracts::{
-		BridgeContractCounterparty, BridgeContractCounterpartyError,
-		BridgeContractCounterpartyResult,
-	},
-	types::{
-		Amount, BridgeTransferDetails, BridgeTransferId, HashLock, HashLockPreImage,
-		InitiatorAddress, RecipientAddress, TimeLock,
-	},
+	counterparty_contract::{CounterpartyCall, SmartContractCounterparty},
+	initiator_contract::{InitiatorCall, SmartContractInitiator},
+	types::{Amount, BridgeAddressType, BridgeHashType, GenUniqueHash, RecipientAddress},
 };
-use rand::prelude::*;
-use serde::Serialize;
-use std::process::Stdio;
-use std::str::FromStr;
-use std::sync::{Arc, RwLock};
-use tokio::{
-	io::{AsyncBufReadExt, BufReader},
-	process::Command as TokioCommand,
-	sync::oneshot,
-	task,
-};
+use event_types::MovementChainEvent;
+use futures::{channel::mpsc, task::AtomicWaker};
 
-use url::Url;
-
-mod event_monitoring;
-mod event_types;
-mod types;
+pub mod client;
+pub mod event_monitoring;
+pub mod event_types;
+pub mod types;
 pub mod utils;
 
-const DUMMY_ADDRESS: AccountAddress = AccountAddress::new([0; 32]);
-const COUNTERPARTY_MODULE_NAME: &str = "atomic_bridge_counterparty";
-
-#[allow(dead_code)]
-enum Call {
-	Lock,
-	Complete,
-	Abort,
-	GetDetails,
+#[derive(Debug)]
+pub enum Transaction<A, H> {
+	Initiator(InitiatorCall<A, H>),
+	Counterparty(CounterpartyCall<A, H>),
 }
 
-pub struct Config {
-	pub rpc_url: Option<String>,
-	pub ws_url: Option<String>,
-	pub chain_id: String,
-	pub signer_private_key: Arc<RwLock<LocalAccount>>,
-	pub initiator_contract: Option<MovementAddress>,
-	pub gas_limit: u64,
+#[allow(unused)]
+pub struct MovementChain<A, H, R> {
+	pub name: String,
+	pub time: u64,
+	pub accounts: HashMap<A, Amount>,
+	pub events: Vec<MovementChainEvent<A, H>>,
+
+	pub initiator_contract: SmartContractInitiator<A, H, R>,
+	pub counterparty_contract: SmartContractCounterparty<A, H>,
+
+	pub transaction_sender: mpsc::UnboundedSender<Transaction<A, H>>,
+	pub transaction_receiver: mpsc::UnboundedReceiver<Transaction<A, H>>,
+
+	pub event_listeners: Vec<mpsc::UnboundedSender<MovementChainEvent<A, H>>>,
+
+	waker: AtomicWaker,
+
+	pub _phantom: std::marker::PhantomData<H>,
 }
 
-impl Config {
-	pub fn build_for_test() -> Self {
-		let seed = [3u8; 32];
-		let mut rng = rand::rngs::StdRng::from_seed(seed);
+impl<A, H, R> MovementChain<A, H, R>
+where
+	A: BridgeAddressType + From<RecipientAddress<A>>,
+	H: BridgeHashType + GenUniqueHash,
+	R: RngSeededClone,
+	H: From<HashLockPreImage>,
+{
+	pub fn new(mut rng: R, name: impl Into<String>) -> Self {
+		let accounts = HashMap::new();
+		let events = Vec::new();
+		let (event_sender, event_receiver) = mpsc::unbounded();
+		let event_listeners = Vec::new();
 
-		Config {
-			rpc_url: Some("http://localhost:8080".parse().unwrap()),
-			ws_url: Some("ws://localhost:8080".parse().unwrap()),
-			chain_id: 4.to_string(),
-			signer_private_key: Arc::new(RwLock::new(LocalAccount::generate(&mut rng))),
-			initiator_contract: None,
-			gas_limit: 10_000_000_000,
-		}
-	}
-}
-
-#[allow(dead_code)]
-#[derive(Clone)]
-pub struct MovementClient {
-	///Address of the counterparty moduke
-	counterparty_address: AccountAddress,
-	///Address of the initiator module
-	initiator_address: Vec<u8>,
-	///The Apotos Rest Client
-	pub rest_client: Client,
-	///The Apotos Rest Client
-	pub faucet_client: Option<Arc<RwLock<FaucetClient>>>,
-	///The signer account
-	signer: Arc<LocalAccount>,
-}
-
-impl MovementClient {
-	pub async fn new(_config: Config) -> Result<Self, anyhow::Error> {
-		let node_connection_url = "http://127.0.0.1:8080".to_string();
-		let node_connection_url = Url::from_str(node_connection_url.as_str()).unwrap();
-
-		let rest_client = Client::new(node_connection_url.clone());
-
-		let seed = [3u8; 32];
-		let mut rng = rand::rngs::StdRng::from_seed(seed);
-		let signer = LocalAccount::generate(&mut rng);
-
-		Ok(MovementClient {
-			counterparty_address: DUMMY_ADDRESS,
-			initiator_address: Vec::new(), //dummy for now
-			rest_client,
-			faucet_client: None,
-			signer: Arc::new(signer),
-		})
-	}
-
-	pub async fn new_for_test(
-		_config: Config,
-	) -> Result<(Self, tokio::process::Child), anyhow::Error> {
-		let (setup_complete_tx, setup_complete_rx) = oneshot::channel();
-		let mut child = TokioCommand::new("aptos")
-			.args(["node", "run-local-testnet"])
-			.stdout(Stdio::piped())
-			.stderr(Stdio::piped())
-			.spawn()?;
-
-		let stdout = child.stdout.take().expect("Failed to capture stdout");
-		let stderr = child.stderr.take().expect("Failed to capture stderr");
-
-		task::spawn(async move {
-			let mut stdout_reader = BufReader::new(stdout).lines();
-			let mut stderr_reader = BufReader::new(stderr).lines();
-
-			loop {
-				tokio::select! {
-					line = stdout_reader.next_line() => {
-						match line {
-							Ok(Some(line)) => {
-								println!("STDOUT: {}", line);
-								if line.contains("Setup is complete") {
-									println!("Testnet is up and running!");
-									let _ = setup_complete_tx.send(());
-																	return Ok(());
-								}
-							},
-							Ok(None) => {
-								return Err(anyhow::anyhow!("Unexpected end of stdout stream"));
-							},
-							Err(e) => {
-								return Err(anyhow::anyhow!("Error reading stdout: {}", e));
-							}
-						}
-					},
-					line = stderr_reader.next_line() => {
-						match line {
-							Ok(Some(line)) => {
-								println!("STDERR: {}", line);
-								if line.contains("Setup is complete") {
-									println!("Testnet is up and running!");
-									let _ = setup_complete_tx.send(());
-																	return Ok(());
-								}
-							},
-							Ok(None) => {
-								return Err(anyhow::anyhow!("Unexpected end of stderr stream"));
-							}
-							Err(e) => {
-								return Err(anyhow::anyhow!("Error reading stderr: {}", e));
-							}
-						}
-					}
-				}
-			}
-		});
-
-		setup_complete_rx.await.expect("Failed to receive setup completion signal");
-		println!("Setup complete message received.");
-
-		let node_connection_url = "http://127.0.0.1:8080".to_string();
-		let node_connection_url = Url::from_str(node_connection_url.as_str()).unwrap();
-		let rest_client = Client::new(node_connection_url.clone());
-
-		let faucet_url = "http://127.0.0.1:8081".to_string();
-		let faucet_url = Url::from_str(faucet_url.as_str()).unwrap();
-		let faucet_client = Arc::new(RwLock::new(FaucetClient::new(
-			faucet_url.clone(),
-			node_connection_url.clone(),
-		)));
-
-		let mut rng = ::rand::rngs::StdRng::from_seed([3u8; 32]);
-		Ok((
-			MovementClient {
-				counterparty_address: DUMMY_ADDRESS,
-				initiator_address: Vec::new(), // dummy for now
-				rest_client,
-				faucet_client: Some(faucet_client),
-				signer: Arc::new(LocalAccount::generate(&mut rng)),
-			},
-			child,
-		))
-	}
-
-	pub fn rest_client(&self) -> &Client {
-		&self.rest_client
-	}
-
-	pub fn faucet_client(&self) -> Result<&Arc<RwLock<FaucetClient>>> {
-		if let Some(faucet_client) = &self.faucet_client {
-			Ok(faucet_client)
-		} else {
-			Err(anyhow::anyhow!("Faucet client not initialized"))
-		}
-	}
-}
-
-#[async_trait::async_trait]
-impl BridgeContractCounterparty for MovementClient {
-	type Address = MovementAddress;
-	type Hash = [u8; 32];
-
-	async fn lock_bridge_transfer_assets(
-		&mut self,
-		bridge_transfer_id: BridgeTransferId<Self::Hash>,
-		hash_lock: HashLock<Self::Hash>,
-		time_lock: TimeLock,
-		initiator: InitiatorAddress<Vec<u8>>,
-		recipient: RecipientAddress<Self::Address>,
-		amount: Amount,
-	) -> BridgeContractCounterpartyResult<()> {
-		//@TODO properly return an error instead of unwrapping
-		let args = vec![
-			bcs::to_bytes(&initiator.0).unwrap(),
-			bcs::to_bytes(&bridge_transfer_id.0).unwrap(),
-			bcs::to_bytes(&hash_lock.0).unwrap(),
-			bcs::to_bytes(&time_lock.0).unwrap(),
-			bcs::to_bytes(&recipient.0).unwrap(),
-			bcs::to_bytes(&amount.0).unwrap(),
-		];
-		let payload = utils::make_aptos_payload(
-			self.counterparty_address,
-			COUNTERPARTY_MODULE_NAME,
-			"lock_bridge_transfer_assets",
-			self.counterparty_type_args(Call::Lock),
-			args,
-		);
-		let _ = utils::send_and_confirm_aptos_transaction(
-			&self.rest_client,
-			self.signer.as_ref(),
-			payload,
-		)
-		.await
-		.map_err(|_| BridgeContractCounterpartyError::LockTransferAssetsError);
-		Ok(())
-	}
-
-	async fn complete_bridge_transfer(
-		&mut self,
-		bridge_transfer_id: BridgeTransferId<Self::Hash>,
-		preimage: HashLockPreImage,
-	) -> BridgeContractCounterpartyResult<()> {
-		let args = vec![
-			bcs::to_bytes(&self.signer.address()).unwrap(),
-			bcs::to_bytes(&bridge_transfer_id.0).unwrap(),
-			bcs::to_bytes(&preimage.0).unwrap(),
-		];
-		let payload = utils::make_aptos_payload(
-			self.counterparty_address,
-			COUNTERPARTY_MODULE_NAME,
-			"complete_bridge_transfer",
-			self.counterparty_type_args(Call::Complete),
-			args,
-		);
-
-		let _ = utils::send_and_confirm_aptos_transaction(
-			&self.rest_client,
-			self.signer.as_ref(),
-			payload,
-		)
-		.await
-		.map_err(|_| BridgeContractCounterpartyError::CompleteTransferError);
-		Ok(())
-	}
-
-	async fn abort_bridge_transfer(
-		&mut self,
-		bridge_transfer_id: BridgeTransferId<Self::Hash>,
-	) -> BridgeContractCounterpartyResult<()> {
-		let args = vec![
-			bcs::to_bytes(&self.signer.address()).unwrap(),
-			bcs::to_bytes(&bridge_transfer_id.0).unwrap(),
-		];
-		let payload = utils::make_aptos_payload(
-			self.counterparty_address,
-			COUNTERPARTY_MODULE_NAME,
-			"abort_bridge_transfer",
-			self.counterparty_type_args(Call::Abort),
-			args,
-		);
-		let _ = utils::send_and_confirm_aptos_transaction(
-			&self.rest_client,
-			self.signer.as_ref(),
-			payload,
-		)
-		.await
-		.map_err(|_| BridgeContractCounterpartyError::AbortTransferError);
-		Ok(())
-	}
-
-	async fn get_bridge_transfer_details(
-		&mut self,
-		_bridge_transfer_id: BridgeTransferId<Self::Hash>,
-	) -> BridgeContractCounterpartyResult<Option<BridgeTransferDetails<Self::Address, Self::Hash>>>
-	{
-		todo!();
-	}
-}
-
-impl MovementClient {
-	fn counterparty_type_args(&self, call: Call) -> Vec<TypeTag> {
-		match call {
-			Call::Lock => vec![TypeTag::Address, TypeTag::U64, TypeTag::U64, TypeTag::U8],
-			Call::Complete => vec![TypeTag::Address, TypeTag::U64, TypeTag::U8],
-			Call::Abort => vec![TypeTag::Address, TypeTag::U64],
-			Call::GetDetails => vec![TypeTag::Address, TypeTag::U64],
+		Self {
+			name: name.into(),
+			time: 0,
+			accounts,
+			events,
+			initiator_contract: SmartContractInitiator::new(rng.seeded_clone()),
+			counterparty_contract: SmartContractCounterparty::new(),
+			transaction_sender: event_sender,
+			transaction_receiver: event_receiver,
+			event_listeners,
+			waker: AtomicWaker::new(),
+			_phantom: std::marker::PhantomData,
 		}
 	}
 }
